+++ src/samplers/sobol.rs
@@ -158,7 +158,7 @@ impl SobolSampler {
             let n_samples = self.samples_1d_array_sizes[i] * self.samples_per_pixel as i32;
             for j in 0..n_samples {
                 let index: u64 = self.get_index_for_sample(j as u64);
-                self.sample_array_1d[i as usize][j as usize] =
+                self.sample_array_1d[i][j as usize] =
                     self.sample_dimension(index, self.array_start_dim + i as i64);
             }
         }
